requirements.txt
aiogram>=3.3.0
celery>=5.3.6
redis>=5.0.1
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
boto3>=1.34.14
tenacity>=8.2.3
httpx>=0.26.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
sentry-sdk>=1.39.2
yt-dlp>=2023.12.30
openai>=1.10.0